Add Makefile, tests, Dockerfile and update CI pipeline
- Add Makefile with all/test/run/clean targets
- Add tests/test_twosum.cpp with 24 dedicated unit tests (PASS/FAIL style)
- Simplify ci.yml to ubuntu-only pipeline using make all + make test
- Add Dockerfile using gcc:14 image, builds and runs tests on container start
- Add .gitignore to exclude build/, *.exe, mingw64/, and temp files
- Fix sign-compare warning in src/twosum.cpp

+++ src/twosum.cpp
@@ -25,7 +25,7 @@ std::vector<int> twoSumArray(const std::vector<int>& nums, int target) {
 std::vector<int> twoSumHashTable(const std::vector<int>& nums, int target) {
     std::unordered_map<int, int> value_to_index;
     
-    for (int i = 0; i < nums.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(nums.size()); ++i) {
         int complement = target - nums[i];
         
         // Check if complement exists in the map
